# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Peminjaman",
        "description": "Dokumentasi API untuk fitur peminjaman buku",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/peminjaman": {
            "get": {
                "tags": [
                    "Peminjaman"
                ],
                "summary": "Ambil semua data peminjaman",
                "operationId": "4adb481982e6d55fa909c818eb236673",
                "responses": {
                    "200": {
                        "description": "Data berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Peminjaman"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Peminjaman"
                ],
                "summary": "Tambah data peminjaman baru",
                "operationId": "efd17f3ef59574270e1c8c4558173095",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "book_id",
                                    "tanggal_pinjam",
                                    "tanggal_kembali",
                                    "status"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "book_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "tanggal_pinjam": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-05-20"
                                    },
                                    "tanggal_kembali": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-05-27"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "dipinjam"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Peminjaman"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/api/peminjaman/{id}": {
            "get": {
                "tags": [
                    "Peminjaman"
                ],
                "summary": "Ambil data peminjaman berdasarkan ID",
                "operationId": "df8a2af2f70560e288c7dbcf8e17c25c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Peminjaman"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Peminjaman"
                ],
                "summary": "Update data peminjaman berdasarkan ID",
                "operationId": "544066d258653cd42ea67d3dee490114",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "book_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "tanggal_pinjam": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-05-20"
                                    },
                                    "tanggal_kembali": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-05-27"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "dikembalikan"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Peminjaman"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Peminjaman"
                ],
                "summary": "Hapus data peminjaman berdasarkan ID",
                "operationId": "5d908e88d0b4a1854835c1c1b8e58f62",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Data berhasil dihapus"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Peminjaman": {
                "title": "Peminjaman",
                "required": [
                    "user_id",
                    "book_id",
                    "tanggal_pinjam",
                    "tanggal_kembali",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "book_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "tanggal_pinjam": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-05-20"
                    },
                    "tanggal_kembali": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-05-27"
                    },
                    "status": {
                        "type": "string",
                        "example": "dipinjam"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}